Address feedback from internal mtg
