+++ examples/config_example.json
@@ -3,7 +3,7 @@
   "categorical_parameter_1": "option1",
   "float_parameter_1": 1.0,
   "list_parameter_1": [0, 1, 2],
-  "loss": {
+  "nested_section": {
     "str_parameter_1": null,
     "int_parameter_1": 1
   }
