+++ examples/config_example.yaml
@@ -1,0 +1,7 @@
+boolean_parameter_1: true
+categorical_parameter_1: "option1"
+float_parameter_1: 1.0
+list_parameter_1: [0, 1, 2]
+nested_section:
+    str_parameter_1: null
+    int_parameter_1: 1
