--- sample/nested_config_values.json
@@ -1,31 +1,0 @@
-{
-  "model": {
-    "type": "boosted_tree",
-    "parameters": {
-      "max_depth": 5,
-      "max_leaves": 31,
-      "bagging_fraction": 0.5,
-      "eta": 0.03,
-      "lr": 0.01
-    }
-  },
-  "train": {
-    "data": {
-      "paths": ["data/00*.csv", "data/01*.csv"],
-      "augmentation": [
-        {
-          "method": "fillna",
-          "column": "x",
-          "kwargs": {
-            "value": -1
-          }
-        }
-      ]
-    }
-  },
-  "eval": {
-    "data": {
-      "paths": ["data/05*.csv"]
-    }
-  }
-}
